projects/01-itgc-evidence-analyser/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
click>=8.1.0
pdfplumber>=0.11.0
reportlab>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
Pillow>=10.0.0
pyyaml>=6.0
rich>=13.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
